rename: shuttle을 reservation으로 변경
