+++ src/app/reservation/[id]/write/components/ShuttleWriteStep2.tsx
