+++ src/app/reservation/[id]/write/sections/PriceDetail.tsx
