+++ src/app/reservation/[id]/write/sections/ReservationCompleted.tsx
