+++ src/app/reservation/[id]/write/sections/ReservationInfo.tsx
