+++ src/app/reservation/[id]/write/sections/TotalPriceInfo.tsx
